requirements.txt
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.5
djangorestframework==3.14.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytz==2023.3.post1
replicate==0.15.3
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
